refactor: split config.go into config_gateways.go and config_validate.go
Split the 1000+ line god file into focused files:
- config.go: core Config struct, providers, runtime, storage, load/save
- config_gateways.go: GatewaysConfig + 30 platform structs + marshal/unmarshal
- config_validate.go: Validate + FillDefaults

No behavior changes — pure file reorganization.
